comment out requiring common library in hdtype, create lib/entities.olmec.lua, move olmec methods into olmec.lua
