+++ lib/entities/hdtype.lua
@@ -1,4 +1,4 @@
-commonlib = require 'common'
+-- commonlib = require 'common'
 
 local module = {}
